Massive refactor 1/n
